+++ packages/lib-sourcify/test/Compilation/VyperCompilation.spec.ts
@@ -647,4 +647,44 @@ describe('VyperCompilation', () => {
       version: 1,
     });
   });
+
+  it('should clean compiler version with v prefix', () => {
+    const contractPath = path.join(
+      __dirname,
+      '..',
+      'sources',
+      'Vyper',
+      'testcontract',
+    );
+    const contractFileName = 'test.vy';
+    const contractContent = fs.readFileSync(
+      path.join(contractPath, contractFileName),
+      'utf8',
+    );
+
+    const compilation = new VyperCompilation(
+      vyperCompiler,
+      'v0.3.10+commit.91361694',
+      {
+        language: 'Vyper',
+        sources: {
+          [contractFileName]: {
+            content: contractContent,
+          },
+        },
+        settings: {
+          evmVersion: 'istanbul',
+          outputSelection: {
+            '*': ['evm.bytecode'],
+          },
+        },
+      },
+      {
+        name: contractFileName.split('.')[0],
+        path: contractFileName,
+      },
+    );
+
+    expect(compilation.compilerVersion).to.equal('0.3.10+commit.91361694');
+  });
 });